SmartDeskAI_Assistant/requirements.txt
# Core audio and data processing
numpy
sounddevice
soundfile

# Display and imaging
Pillow
luma.oled

# Raspberry Pi hardware access
spidev
rpi-lgpio
gpiozero

# Offline speech input
faster-whisper
silero-vad
onnxruntime

# Offline speech output
piper

# Test runner
pytest

# Optional semantic intent upgrade
# Install this only if you want embedding-based semantic matching on the Pi.
# The assistant can still run without it because the matcher degrades gracefully.
# sentence-transformers